+++ actionview/lib/action_view/template/resolver.rb
@@ -164,20 +164,15 @@ module ActionView
     EXTENSIONS = { locale: ".", formats: ".", variants: "+", handlers: "." }
     DEFAULT_PATTERN = ":prefix/:action{.:locale,}{.:formats,}{+:variants,}{.:handlers,}"
 
-    def initialize(pattern = nil)
-      if pattern
-        ActiveSupport::Deprecation.warn "Specifying a custom path for #{self.class} is deprecated. Implement a custom Resolver subclass instead."
-        @pattern = pattern
-      else
-        @pattern = DEFAULT_PATTERN
-      end
+    def initialize
+      @pattern = DEFAULT_PATTERN
       @unbound_templates = Concurrent::Map.new
-      super()
+      super
     end
 
     def clear_cache
       @unbound_templates.clear
-      super()
+      super
     end
 
     private
@@ -300,9 +295,9 @@ module ActionView
   class FileSystemResolver < PathResolver
     attr_reader :path
 
-    def initialize(path, pattern = nil)
+    def initialize(path)
       raise ArgumentError, "path already is a Resolver class" if path.is_a?(Resolver)
-      super(pattern)
+      super()
       @path = File.expand_path(path)
     end